refactor(threads): split actions and thread item hotspots
