+++ docs/architecture/large-file-near-threshold-watchlist.md
@@ -1,26 +1,41 @@
-# Large File Baseline
+# Large File Near-Threshold Watchlist
 
-- Generated at: 2026-03-16T05:25:48.162Z
-- Threshold: > 2500 lines
-- Count: 17
+- Generated at: 2026-04-22T17:01:09.717Z
+- Scope: warn
+- Policy version: 2026-04-22.policy-v2
+- Count: 31
 
-| File | Lines | Type | Priority |
-|---|---:|---|---|
-| `src/features/settings/components/SettingsView.tsx` | 2995 | ts/tsx | P0 |
-| `src-tauri/src/engine/claude.rs` | 2985 | rust | P0 |
-| `src/app-shell.tsx` | 2971 | ts/tsx | P1 |
-| `src/styles/settings.part2.css` | 2957 | css | P1 |
-| `src-tauri/src/backend/app_server.rs` | 2925 | rust | P0 |
-| `src/features/messages/components/Messages.test.tsx` | 2869 | ts/tsx | P1 |
-| `src/features/threads/hooks/useThreadsReducer.ts` | 2814 | ts/tsx | P1 |
-| `src/styles/settings.part1.css` | 2807 | css | P1 |
-| `src/styles/git-history.part1.css` | 2800 | css | P1 |
-| `src/styles/git-history.part2.css` | 2776 | css | P1 |
-| `src/features/git-history/components/git-history-panel/components/GitHistoryPanelImpl.tsx` | 2767 | ts/tsx | P0 |
-| `src-tauri/src/engine/commands.rs` | 2743 | rust | P0 |
-| `src/features/git/components/GitDiffPanel.tsx` | 2723 | ts/tsx | P1 |
-| `src/styles/spec-hub.css` | 2603 | css | P1 |
-| `src/features/messages/components/Messages.tsx` | 2577 | ts/tsx | P1 |
-| `src-tauri/src/git/commands.rs` | 2547 | rust | P0 |
-| `src/features/threads/hooks/useThreadMessaging.ts` | 2537 | ts/tsx | P1 |
+| File | Lines | Type | Policy | Priority | Warn | Fail | Severity | Status | Baseline | Delta |
+|---|---:|---|---|---|---:|---:|---|---|---:|---:|
+| `src/features/threads/hooks/useThreadActions.test.tsx` | 2995 | ts/tsx | test-files | P2 | 2600 | 3000 | warn | watch |  |  |
+| `src/styles/settings.part1.css` | 2993 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
+| `src-tauri/src/engine/commands.rs` | 2976 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | fail | captured |  |  |
+| `src/features/app/hooks/useAppServerEvents.test.tsx` | 2966 | ts/tsx | test-files | P2 | 2600 | 3000 | warn | watch |  |  |
+| `src/features/threads/hooks/useThreadsReducer.test.ts` | 2961 | ts/tsx | test-files | P2 | 2600 | 3000 | warn | watch |  |  |
+| `src/styles/composer.part1.css` | 2941 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
+| `src-tauri/src/local_usage.rs` | 2929 | rust | default-source | P1 | 2600 | 3000 | warn | watch |  |  |
+| `src/features/messages/components/Messages.test.tsx` | 2926 | ts/tsx | test-files | P2 | 2600 | 3000 | warn | watch |  |  |
+| `src-tauri/src/bin/cc_gui_daemon.rs` | 2882 | rust | default-source | P1 | 2600 | 3000 | warn | watch |  |  |
+| `src/styles/settings.part2.css` | 2860 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
+| `src/i18n/locales/en.part1.ts` | 2835 | i18n | i18n | P2 | 2600 | 3000 | warn | watch |  |  |
+| `src/styles/git-history.part1.css` | 2818 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
+| `src/features/git/components/GitDiffPanel.tsx` | 2807 | ts/tsx | default-source | P1 | 2600 | 3000 | warn | watch |  |  |
+| `src/styles/git-history.part2.css` | 2776 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
+| `src/features/git-history/components/git-history-panel/components/GitHistoryPanelImpl.tsx` | 2773 | ts/tsx | feature-hotpath | P1 | 2400 | 2800 | warn | watch |  |  |
+| `src/i18n/locales/zh.part1.ts` | 2743 | i18n | i18n | P2 | 2600 | 3000 | warn | watch |  |  |
+| `src/features/threads/hooks/useThreadsReducer.ts` | 2723 | ts/tsx | feature-hotpath | P1 | 2400 | 2800 | warn | watch |  |  |
+| `src-tauri/src/bin/cc_gui_daemon/daemon_state.rs` | 2713 | rust | default-source | P1 | 2600 | 3000 | warn | watch |  |  |
+| `src-tauri/src/runtime/mod.rs` | 2655 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | fail | captured |  |  |
+| `src/styles/sidebar.css` | 2640 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
+| `src-tauri/src/git/commands.rs` | 2619 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | fail | captured |  |  |
+| `src/styles/spec-hub.css` | 2609 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
+| `src/features/settings/components/SettingsView.tsx` | 2607 | ts/tsx | feature-hotpath | P1 | 2400 | 2800 | warn | watch |  |  |
+| `src-tauri/src/codex/mod.rs` | 2574 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | warn | watch |  |  |
+| `src/services/tauri.ts` | 2567 | ts/tsx | bridge-runtime-critical | P0 | 2200 | 2600 | warn | watch |  |  |
+| `src/app-shell.tsx` | 2509 | ts/tsx | bridge-runtime-critical | P0 | 2200 | 2600 | warn | watch |  |  |
+| `src/styles/tool-blocks.css` | 2265 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
+| `src/styles/file-view-panel.css` | 2237 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
+| `src-tauri/src/backend/app_server.rs` | 2235 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | warn | watch |  |  |
+| `src/styles/messages.part1.css` | 2208 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
+| `src-tauri/src/engine/gemini.rs` | 2205 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | warn | watch |  |  |
