+++ openspec/changes/split-thread-items-assistant-text-normalization/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-04-22
